+++ StripePaymentSheet/StripePaymentSheet/Source/PaymentSheet/PaymentSheet.swift
@@ -59,6 +59,15 @@ public class PaymentSheet {
             }
         }
 
+        var checkout: Checkout? {
+            switch self {
+            case .checkout(let checkout):
+                return checkout
+            case .paymentIntentClientSecret, .setupIntentClientSecret, .deferredIntent:
+                return nil
+            }
+        }
+
         var isDeferred: Bool {
             if case .deferredIntent = self {
                 return true
@@ -324,12 +333,7 @@ extension PaymentSheet: PaymentSheetViewControllerDelegate {
                 paymentOption: paymentOption,
                 paymentHandler: self.paymentHandler,
                 integrationShape: .complete,
-                checkout: {
-                    guard case .checkout(let checkout) = self.mode else {
-                        return nil
-                    }
-                    return checkout
-                }(),
+                checkout: self.mode.checkout,
                 confirmationChallenge: self.confirmationChallenge,
                 analyticsHelper: self.analyticsHelper
             ) { result, deferredIntentConfirmationType in