+++ StripePaymentSheet/StripePaymentSheetTests/PaymentSheet/PaymentSheetFormFactoryTest.swift
@@ -15,19 +15,13 @@ import XCTest
 @testable@_spi(STP) import StripePaymentsUI
 @testable@_spi(STP) import StripeUICore
 
-class MockElement: Element {
+private final class ParamsCountingPaymentMethodElement: PaymentMethodElement {
     var collectsUserInput: Bool = false
-
-    var paramsUpdater: (IntentConfirmParams) -> IntentConfirmParams?
-
-    init(
-        paramsUpdater: @escaping (IntentConfirmParams) -> IntentConfirmParams?
-    ) {
-        self.paramsUpdater = paramsUpdater
-    }
+    var updateParamsCallCount = 0
 
     func updateParams(params: IntentConfirmParams) -> IntentConfirmParams? {
-        return paramsUpdater(params)
+        updateParamsCallCount += 1
+        return params
     }
 
     weak var delegate: ElementDelegate?
@@ -101,6 +95,68 @@ class PaymentSheetFormFactoryTest: XCTestCase {
         XCTAssertEqual(params?.paymentMethodParams.billingDetails?.email, "email@stripe.com")
         XCTAssertEqual(params?.paymentMethodParams.type, .SEPADebit)
         XCTAssertEqual(params?.paymentMethodType, .stripe(.SEPADebit))
+    }
+
+    func testPaymentMethodElementWrapperDoesNotImplicitlyUpdateParamsFromWrappedElement() {
+        // Given
+        let element = ParamsCountingPaymentMethodElement()
+        let wrapper = PaymentMethodElementWrapper(element) { _, params in params }
+
+        // When
+        _ = wrapper.updateParams(params: IntentConfirmParams(type: .stripe(.card)))
+
+        // Then
+        XCTAssertEqual(element.updateParamsCallCount, 0)
+    }
+
+    func testPaymentMethodElementWrapperExplicitlyUpdatesParamsFromWrappedElementOnce() {
+        // Given
+        let element = ParamsCountingPaymentMethodElement()
+        let wrapper = PaymentMethodElementWrapper(updatingParamsFrom: element) { _, params in params }
+
+        // When
+        _ = wrapper.updateParams(params: IntentConfirmParams(type: .stripe(.card)))
+
+        // Then
+        XCTAssertEqual(element.updateParamsCallCount, 1)
+    }
+
+    func testPaymentMethodElementWrapperUpdatingParamsFromTextFieldStillValidates() {
+        // Given
+        let textField = TextFieldElement(
+            configuration: TextFieldElement.PANConfiguration(),
+            theme: .default
+        )
+        var paramsUpdaterCallCount = 0
+        let wrapper = PaymentMethodElementWrapper(updatingParamsFrom: textField) { _, params in
+            paramsUpdaterCallCount += 1
+            return params
+        }
+
+        // When
+        let params = wrapper.updateParams(params: IntentConfirmParams(type: .stripe(.card)))
+
+        // Then
+        XCTAssertNil(params)
+        XCTAssertEqual(paramsUpdaterCallCount, 0)
+    }
+
+    func testDefaultsApplierWrapperUpdatesParamsFromWrappedElementOnce() {
+        // Given
+        let factory = PaymentSheetFormFactory(
+            intent: ._testValue(),
+            elementsSession: ._testCardValue(),
+            configuration: .paymentElement(PaymentSheet.Configuration()),
+            paymentMethod: .stripe(.card)
+        )
+        let element = ParamsCountingPaymentMethodElement()
+        let wrapper = factory.makeDefaultsApplierWrapper(for: element)
+
+        // When
+        _ = wrapper.updateParams(params: IntentConfirmParams(type: .stripe(.card)))
+
+        // Then
+        XCTAssertEqual(element.updateParamsCallCount, 1)
     }
 
     func testNameOverrideApiPathBySpec() {